separates opengl into a static library, moves heavy part of shader compilation into asset_kitchen
